+++ src/liberator_mixin/context/core.clj
@@ -1,0 +1,4 @@
+(ns liberator-mixin.context.core)
+
+(defn with-attribute-in-context [key value]
+  {:initialize-context (fn [_] {key value})})